# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Web3.0 区块链系统 API",
    "description": "这是一个使用Golang实现的Web3.0区块链系统的API服务",
    "contact": {
      "name": "API Support",
      "url": "http://www.web3blockchain.io/support",
      "email": "support@web3blockchain.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "用户登录并获取JWT令牌",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "登录信息",
            "name": "login",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.TokenResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        },
        "tags": [
          "认证"
        ],
        "summary": "用户登录"
      }
    },
    "/auth/register": {
      "post": {
        "description": "注册新用户",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "用户注册信息",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.RegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/main.UserResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        },
        "tags": [
          "认证"
        ],
        "summary": "用户注册"
      }
    },
    "/blocks": {
      "get": {
        "description": "返回区块链中的所有区块",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "description": "限制返回的区块数量",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "跳过前面的区块数量",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/main.BlockResponse"
              }
            }
          }
        },
        "tags": [
          "区块"
        ],
        "summary": "获取所有区块"
      }
    },
    "/blocks/{hash}": {
      "get": {
        "description": "通过区块哈希获取区块详情",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "description": "区块哈希",
            "name": "hash",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.BlockResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        },
        "tags": [
          "区块"
        ],
        "summary": "获取特定区块"
      }
    },
    "/info": {
      "get": {
        "description": "返回区块链的基本信息，如区块高度、交易总数等",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "tags": [
          "区块链信息"
        ],
        "summary": "获取区块链信息"
      }
    },
    "/network/config": {
      "get": {
        "description": "获取区块链网络配置参数",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.NetworkConfig"
            }
          }
        },
        "tags": [
          "网络"
        ],
        "summary": "获取网络配置",
        "security": [
          {
            "JWTAuth": []
          }
        ]
      },
      "put": {
        "description": "更新区块链网络配置参数",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "网络配置参数",
            "name": "config",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.NetworkConfig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.NetworkConfig"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        },
        "tags": [
          "网络"
        ],
        "summary": "更新网络配置",
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
    "/transactions": {
      "get": {
        "description": "返回区块链中的所有交易",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "description": "限制返回的交易数量",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "跳过前面的交易数量",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/main.TransactionOutput"
              }
            }
          }
        },
        "tags": [
          "交易"
        ],
        "summary": "获取所有交易"
      },
      "post": {
        "description": "创建一个新交易并广播到区块链网络",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "交易详情",
            "name": "transaction",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.TransactionRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        },
        "tags": [
          "交易"
        ],
        "summary": "创建新交易"
      }
    },
    "/transactions/{id}": {
      "get": {
        "description": "通过交易ID获取交易详情",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "description": "交易ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.TransactionOutput"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        },
        "tags": [
          "交易"
        ],
        "summary": "获取特定交易"
      }
    },
    "/wallets": {
      "get": {
        "description": "获取所有钱包地址列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "tags": [
          "钱包"
        ],
        "summary": "获取所有钱包"
      },
      "post": {
        "description": "创建一个新钱包",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/main.WalletResponse"
            }
          }
        },
        "tags": [
          "钱包"
        ],
        "summary": "创建新钱包"
      }
    },
    "/wallets/{address}/balance": {
      "get": {
        "description": "获取指定钱包地址的余额",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "description": "钱包地址",
            "name": "address",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.BalanceResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        },
        "tags": [
          "钱包"
        ],
        "summary": "获取钱包余额"
      }
    },
    "/wallets/{address}/transactions": {
      "get": {
        "description": "获取指定钱包地址的所有交易历史",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "description": "钱包地址",
            "name": "address",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "限制返回的交易数量",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "跳过前面的交易数量",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/main.TransactionOutput"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        },
        "tags": [
          "钱包"
        ],
        "summary": "获取钱包交易历史"
      }
    }
  },
  "securityDefinitions": {
    "JWTAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "main.BalanceResponse": {
      "type": "object",
      "properties": {
        "balance": {
          "type": "integer"
        }
      }
    },
    "main.BlockResponse": {
      "type": "object",
      "properties": {
        "hash": {
          "type": "string"
        },
        "height": {
          "type": "integer"
        },
        "nonce": {
          "type": "integer"
        },
        "prevBlockHash": {
          "type": "string"
        },
        "timestamp": {
          "type": "integer"
        },
        "transactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/main.TransactionOutput"
          }
        }
      }
    },
    "main.TransactionOutput": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "integer"
        },
        "from": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "inputs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "timestamp": {
          "type": "integer"
        },
        "to": {
          "type": "string"
        }
      }
    },
    "main.TransactionRequest": {
      "type": "object",
      "required": [
        "amount",
        "from",
        "to"
      ],
      "properties": {
        "amount": {
          "type": "integer",
          "minimum": 1
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        }
      }
    },
    "main.WalletResponse": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "publicKey": {
          "type": "string"
        }
      }
    },
    "main.LoginRequest": {
      "type": "object",
      "required": [
        "password",
        "username"
      ],
      "properties": {
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "main.NetworkConfig": {
      "type": "object",
      "properties": {
        "blockTime": {
          "type": "integer",
          "description": "区块生成时间间隔(秒)"
        },
        "difficulty": {
          "type": "integer",
          "description": "挖矿难度"
        },
        "miningReward": {
          "type": "integer",
          "description": "挖矿奖励"
        },
        "txFee": {
          "type": "integer",
          "description": "交易费用"
        }
      }
    },
    "main.RegisterRequest": {
      "type": "object",
      "required": [
        "email",
        "password",
        "username"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "minLength": 6
        },
        "username": {
          "type": "string",
          "minLength": 3
        }
      }
    },
    "main.TokenResponse": {
      "type": "object",
      "properties": {
        "expiresAt": {
          "type": "integer"
        },
        "token": {
          "type": "string"
        },
        "userId": {
          "type": "integer"
        }
      }
    },
    "main.UserResponse": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "walletAddress": {
          "type": "string"
        }
      }
    }
  },
  "tags": [
    {
      "description": "用户认证和授权",
      "name": "认证"
    },
    {
      "description": "网络配置和状态",
      "name": "网络"
    },
    {
      "description": "区块链基本信息",
      "name": "区块链信息"
    },
    {
      "description": "区块相关操作",
      "name": "区块"
    },
    {
      "description": "交易相关操作",
      "name": "交易"
    },
    {
      "description": "钱包相关操作",
      "name": "钱包"
    }
  ]
} 